requirements.txt
pandas>=2.0
pyarrow>=14.0
geopandas>=0.14
shapely>=2.0
pyyaml>=6.0
requests>=2.31
